refactor: Merge stubs & registries
